### 2025-05-17 #8245 @HzaCode
Hi @arfon

Thank you very much for the feedback from the pre-review check. I've reviewed all the points raised.

I will now work on addressing the following:

1.  **DOI Issues:** I will carefully go through my `.bib` file to:
    *   Search for and add DOIs for the entries currently listed under `SKIP DOIs`.
    *   Verify and add the suggested DOIs listed under `MISSING DOIs`.
    *   Investigate and correct the `INVALID DOIs`.

2.  **Paper Compilation Error / Affiliation:** I will update the `paper.md` YAML header to include "Independent Developer" as my affiliation to resolve the potential compilation issue.

3.  **License:** I will also double-check the repository's `LICENSE` file to ensure it is an OSI-approved license as required.

I understand about the 'waitlisted' status and the current backlog. Thank you for letting me know and for your team's efforts. I am happy to wait.

I will make these changes in the repository shortly.

Thanks again for your guidance!
