### 2025-05-31 #8245 @HzaCode
Hi @arfon,
Just a friendly follow-up on my submission. I understand the team is very busy.
If it would help lighten the load for finding a suitable editor, I'd be happy to look through the JOSS editor list and suggest a few individuals whose expertise seems like a good match for "Athlytics." Please let me know if this is something you'd find helpful.
