### 2025-06-24 #8245 @HzaCode
Hi @arfon,

Just gently checking in on this submission. I know it's on the waitlist, but was wondering if there might be an update on editor availability.

No rush at all. Thanks for your time!
