### 2025-09-24 #9075 @editorialbot
[REVIEW]: Athlytics: A Computational Framework for Longitudinal Analysis of Exercise Physiology Metrics from Wearable Sensor Data
**Submitting author:** @HzaCode (<a href="http://orcid.org/0009-0009-0171-4578">Zhiang HEZHIANG</a>)
**Repository:** https://github.com/HzaCode/Athlytics
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
