### 2025-10-29 #9075 @juliaferraioli
Thanks @HzaCode -- I have applied the `paused` label and set myself a reminder to check back in with you. Good luck with the rOpenSci review!
