### 2025-11-05 0e5534c
Fix rOpenSci review issues: convert README badges to standard markdown format
- Convert all HTML badges to standard markdown format
- Use direct badge links (GitHub Actions, Codecov) instead of shields.io redirects
- Remove all HTML tags and convert to pure markdown
- Remove --- separators that caused YAML parse errors
- Fix README.md conversion warning in R CMD check

R CMD check now passes with 0 WARNINGS (down from 1 WARNING)
