+++ README.md
@@ -1,86 +1,25 @@
-
-<div align="center">
-
-<img src="analysis_output/logo.png" alt="Athlytics logo" width="220"/>
-
 # Athlytics
-*An offline, reproducible R toolkit for endurance-exercise data: from FIT/TCX/GPX & Strava archives to validated models (ACWR, EF, pa:hr decoupling, PB, exposure), with built-in QC and uncertainty estimation.*
 
-<p>
-📘 <a href="https://hezhiang.com/Athlytics/"><strong>Docs</strong></a> &nbsp;&bull;&nbsp;
-📦 <a href="https://cran.r-project.org/package=Athlytics"><strong>CRAN</strong></a> &nbsp;&bull;&nbsp;
-🐞 <a href="https://github.com/HzaCode/Athlytics/issues"><strong>Issues</strong></a>
-</p>
+[![R-CMD-check](https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml/badge.svg)](https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml)
+[![CRAN Status](https://www.r-pkg.org/badges/version/Athlytics)](https://cran.r-project.org/package=Athlytics)
+[![Codecov](https://codecov.io/gh/HzaCode/Athlytics/branch/main/graph/badge.svg)](https://app.codecov.io/gh/HzaCode/Athlytics)
+[![Project Status](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 
-<p>
-  <a href="https://cran.r-project.org/package=Athlytics">
-    <img src="https://img.shields.io/badge/CRAN-Accepted-blue?style=flat-square" alt="CRAN">
-  </a>
-  <a href="https://cran.r-project.org/web/views/SportsAnalytics.html">
-    <img src="https://img.shields.io/badge/CRAN%20Listed-Sports%20Analytics-orange?style=flat-square" alt="CRAN Listed">
-  </a>
-  <a href="https://cran.r-project.org/package=Athlytics">
-    <img src="https://cranlogs.r-pkg.org/badges/grand-total/Athlytics?style=flat-square" alt="CRAN downloads (total)">
-  </a>
-  <a href="https://hzacode.r-universe.dev/Athlytics">
-    <img src="https://hzacode.r-universe.dev/badges/Athlytics" alt="R-Universe">
-  </a>
-  <a href="https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml">
-    <img src="https://img.shields.io/github/actions/workflow/status/HzaCode/Athlytics/R-CMD-check.yml?style=flat-square&label=R-CMD-check" alt="R-CMD-check">
-  </a>
-  <a href="https://app.codecov.io/gh/HzaCode/Athlytics">
-    <img src="https://img.shields.io/codecov/c/github/HzaCode/Athlytics?style=flat-square" alt="Coverage">
-  </a>
-  <a href="https://opensource.org/licenses/MIT">
-    <img src="https://img.shields.io/badge/License-MIT-yellow.svg?style=flat-square" alt="MIT License">
-  </a>
-  <a href="https://github.com/firefly-cpp/awesome-computational-intelligence-in-sports?tab=readme-ov-file#software-">
-    <img src="https://awesome.re/badge-flat.svg" alt="Awesome list">
-  </a>
-  <a href="https://docs.ropensci.org/pkgcheck/">
-    <img src="https://img.shields.io/badge/rOpenSci-pkgcheck-brightgreen?style=flat-square" alt="rOpenSci pkgcheck">
-  </a>
-  <a href="https://github.com/HzaCode/Athlytics">
[…]
