### 2025-11-05 51471d7
Fix emoji encoding issues in README.md
- Replace emoji bullets with plain text bullets
- Replace emoji arrows/dashes with standard characters
- Ensure proper UTF-8 encoding for cross-platform compatibility
