+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   advanced_features: advanced_features.html
   athlytics_introduction: athlytics_introduction.html
-last_built: 2025-11-05T17:35Z
+last_built: 2025-11-05T17:43Z
 urls:
   reference: https://hezhiang.com/Athlytics/reference
   article: https://hezhiang.com/Athlytics/articles
