### 2025-11-05 48973f3
Add pkgcheck-action workflow for rOpenSci review
