+++ .github/workflows/pkgcheck.yml
@@ -1,0 +1,29 @@
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+name: pkgcheck
+
+jobs:
+  pkgcheck:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: 'release'
+          use-public-rspm: true
+
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          extra-packages: any::rcmdcheck, any::roxygen2, any::pkgbuild
+
+      - uses: ropensci-review-tools/pkgcheck-action@main
+        env:
+          GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+
