### 2025-11-07 a9ae391
Fix GitHub Actions workflows and R CMD check warnings
