+++ .github/workflows/pkgdown.yml
@@ -37,6 +37,9 @@ jobs:
         run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
         shell: Rscript {0}
 
+      - name: Create .nojekyll file
+        run: touch docs/.nojekyll
+
       - name: Deploy to GitHub pages 🚀
         if: github.event_name != 'pull_request'
         uses: JamesIves/github-pages-deploy-action@v4.6.8
