### 2025-11-07 a98fe97
Add FITfileR to Suggests field to fix R CMD check warning
