### 2025-12-09 1ed4d6c
docs: add comprehensive badge collection to README
