### 2025-12-09 8de50d5
docs: add logo back to README with centered layout
