+++ README.md
@@ -1,3 +1,7 @@
+<div align="center">
+
+<img src="man/figures/image.png" alt="Athlytics logo" width="220"/>
+
 # Athlytics
 
 [![CRAN Status](https://img.shields.io/badge/CRAN-Accepted-blue?style=flat-square)](https://cran.r-project.org/package=Athlytics)
@@ -9,6 +13,8 @@
 [![MIT License](https://img.shields.io/badge/License-MIT-yellow.svg?style=flat-square)](https://opensource.org/licenses/MIT)
 [![Awesome](https://awesome.re/badge-flat.svg)](https://github.com/firefly-cpp/awesome-computational-intelligence-in-sports?tab=readme-ov-file#software-)
 [![Project Status](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+
+</div>
 
 ## Overview
 
