+++ articles/index.md
@@ -1,0 +1,11 @@
+# Articles
+
+### Getting Started
+
+- [Getting Started with
+  Athlytics](https://hezhiang.com/Athlytics/articles/athlytics_introduction.md):
+
+### Advanced Usage
+
+- [Advanced Features: EWMA, Quality Control, and Cohort
+  Analysis](https://hezhiang.com/Athlytics/articles/advanced_features.md):
