+++ authors.md
@@ -1,0 +1,22 @@
+# Authors and Citation
+
+## Authors
+
+- **Zhiang He**. Author, maintainer.
+
+## Citation
+
+Source:
+[`inst/CITATION`](https://github.com/HzaCode/Athlytics/blob/main/inst/CITATION)
+
+Zhiang He (2025). Athlytics: A Computational Framework for Longitudinal
+Analysis of Exercise Physiology. R package version 1.0.0.
+https://github.com/HzaCode/Athlytics
+
+    @Manual{,
+      title = {Athlytics: A Computational Framework for Longitudinal Analysis of Exercise Physiology},
+      author = {Zhiang He},
+      year = {2025},
+      note = {R package version 1.0.0},
+      url = {https://github.com/HzaCode/Athlytics},
+    }

[…]
