### 2025-12-11 1a67ea6
Fix plot issues and clean up code
