+++ reference/athlytics_palette_nature.md
@@ -17,9 +17,8 @@ A character vector of 9 hex color codes
 
 ``` r
 if (FALSE) { # \dontrun{
-library(ggplot2)
-ggplot(data, aes(x, y, color = group)) +
-  geom_line() +
-  scale_color_manual(values = athlytics_palette_nature())
+ggplot2::ggplot(data, ggplot2::aes(x, y, color = group)) +
+  ggplot2::geom_line() +
+  ggplot2::scale_color_manual(values = athlytics_palette_nature())
 } # }
 ```

[…]
