### 2025-12-18 64134dd
refactor: reduce cyclomatic complexity and remove unused viridis dependency
- Extract daily load calculation into internal helper (R/internal_load.R)
- Refactor calculate_acwr() to use calculate_daily_load_internal()
- Refactor calculate_exposure() to use the same helper
- Add validate_load_metric_params() for centralized validation
- Remove viridis from Imports (DESCRIPTION, NAMESPACE, codemeta.json)
- Remove @importFrom viridis from plot_exposure.R

Addresses reviewer comments on:
1. High cyclomatic complexity in core functions
2. Unused viridis package dependency

Public API unchanged - no 
[…]
