+++ NAMESPACE
@@ -105,6 +105,5 @@ importFrom(tidyr,pivot_wider)
 importFrom(tidyr,unnest)
 importFrom(tools,toTitleCase)
 importFrom(utils,read.csv)
-importFrom(viridis,scale_color_viridis)
 importFrom(zoo,rollapply)
 importFrom(zoo,rollmean)

[…]
