### 2025-12-18 39ea70c
v1.0.1: reduce complexity, remove viridis, fix notes
