+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://hezhiang.com/Athlytics/
+url: https://hzacode.github.io/Athlytics/
 
 template:
   bootstrap: 5
