+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: Athlytics
 Title: Academic R Package for Sports Physiology Analysis from Local 'Strava' Data
-Version: 1.0.0
+Version: 1.0.1
 Author: Zhiang He [aut, cre]
 Maintainer: Zhiang He <ang@hezhiang.com>
 Authors@R: 
