### 2026-01-03 3c229d4
clean: remove debug messages from calculate_acwr
