+++ llms.txt
@@ -200,9 +200,9 @@ comply with local IRB/ethics and data-protection regulations.
 ## 🤝 Contributing
 
 Contributions are welcome! Please read our
-[CONTRIBUTING.md](https://hezhiang.com/Athlytics/CONTRIBUTING.md) guide
-and follow our [Code of
-Conduct](https://hezhiang.com/Athlytics/CODE_OF_CONDUCT.md).
+[CONTRIBUTING.md](https://hzacode.github.io/Athlytics/CONTRIBUTING.md)
+guide and follow our [Code of
+Conduct](https://hzacode.github.io/Athlytics/CODE_OF_CONDUCT.md).
 
 - **🐛 Report an Issue**: [Open an
   Issue](https://github.com/HzaCode/Athlytics/issues)
@@ -224,102 +224,102 @@ valuable feedback and suggestions.
 
 Functions for loading and parsing local Strava data
 
-- [`load_local_activities()`](https://hezhiang.com/Athlytics/reference/load_local_activities.md)
+- [`load_local_activities()`](https://hzacode.github.io/Athlytics/reference/load_local_activities.md)
   : Load Activities from Local Strava Export
-- [`parse_activity_file()`](https://hezhiang.com/Athlytics/reference/parse_activity_file.md)
+- [`parse_activity_file()`](https://hzacode.github.io/Athlytics/reference/parse_activity_file.md)
   : Parse Activity File (FIT, TCX, or GPX)
 
 ## Core Metrics Calculation
 
 Calculate key physiological metrics
 
-- [`calculate_acwr()`](https://hezhiang.com/Athlytics/reference/calculate_acwr.md)
+- [`calculate_acwr()`](https://hzacode.github.io/Athlytics/reference/calculate_acwr.md)
   : Calculate Acute:Chronic Workload Ratio (ACWR)
-- [`calculate_acwr_ewma()`](https://hezhiang.com/Athlytics/reference/calculate_acwr_ewma.md)
+- [`calculate_acwr_ewma()`](https://hzacode.github.io/Athlytics/reference/calculate_acwr_ewma.md)
   : Calculate ACWR using EWMA Method with Confidence Bands
-- [`calculate_ef()`](https://hezhiang.com/Athlytics/reference/calculate_ef.md)
+- [`calculate_ef()`](https://hzacode.github.io/Athlytics/reference/calculate_ef.md)
   : Calculate Efficiency Factor (EF)
-- [`calculate_ef_from_stream()`](https://hezhiang.com/Athlytics/reference/calculate_ef_from_stream.md)
+- [`calculate_ef_from_stream()`](https://hzacode.github.io/Athlytics/reference/calculate_ef_from_stream.md)
   : Calculate EF from Stream Data with Steady-State Analysis
-- [`calculate_decoupling()`](https://hezhiang.com/Athlytics/reference/calculate_decoupling.md)
+- [`calculate_decoupling()`](https://hzacode.github.io/Athlytics/reference/calculate_decoupling.md)
   : Calculate Aerobic Decoupling
-- [`calculate_exposure()`](https://hezhiang.com/Athlytics/reference/calculate_exposure.md)
+- [`calculate_exposure()`](https://hzacode.github.io/Athlytics/reference/calculate_exposure.md)
   : Calculate Training Load Exposure (ATL, CTL, ACWR)
-- [`calculate_pbs()`](https://hezhiang.com/Athlytics/reference/calculate_pbs.md)
+- [`calculate_pbs()`](https://hzacode.github.io/Athlytics/reference/calculate_pbs.md)
   : Calculate Personal Bests (PBs) from Local Strava Data
 
 ## Visualization
 
 Plot and visualize training metrics
 
[…]
