+++ reference/calculate_pbs.md
@@ -22,7 +22,7 @@ calculate_pbs(
 - activities_data:
 
   A data frame of activities from
-  [`load_local_activities()`](https://hezhiang.com/Athlytics/reference/load_local_activities.md).
+  [`load_local_activities()`](https://hzacode.github.io/Athlytics/reference/load_local_activities.md).
   Must contain columns: date, type, filename, distance.
 
 - export_dir:
