+++ reference/add_reference_bands.md
@@ -24,7 +24,7 @@ add_reference_bands(
 - reference_data:
 
   A data frame from
-  [`cohort_reference()`](https://hezhiang.com/Athlytics/reference/cohort_reference.md).
+  [`cohort_reference()`](https://hzacode.github.io/Athlytics/reference/cohort_reference.md).
 
 - bands:
 
