### 2026-01-03 8e70297
refactor: remove code duplication and fix docs
