### 2026-01-03 8d9dbc3
fix: remove remaining code duplication in calculate_ef
