+++ man/plot_exposure.Rd
@@ -61,7 +61,7 @@ If \code{exposure_df} is not provided, it calls \code{calculate_exposure} first.
 }
 \examples{
 # Example using simulated data
-data(Athlytics_sample_data)
+data(athlytics_sample_exposure)
 # Ensure exposure_df is named and other necessary parameters like activity_type are provided
 p <- plot_exposure(exposure_df = athlytics_sample_exposure, activity_type = "Run")
 print(p)
