### 2026-01-03 7bb6642
fix: correct dataset names in roxygen examples
