+++ R/calculate_decoupling.R
@@ -57,7 +57,7 @@
 #'
 #' @examples
 #' # Example using simulated data
-#' data(Athlytics_sample_data)
+#' data(athlytics_sample_decoupling)
 #' print(head(athlytics_sample_decoupling))
 #'
 #' \dontrun{
