+++ reference/plot_exposure.md
@@ -93,8 +93,7 @@ can be shaded. If `exposure_df` is not provided, it calls
 
 ``` r
 # Example using simulated data
-data(Athlytics_sample_data)
-#> Warning: data set ‘Athlytics_sample_data’ not found
+data(athlytics_sample_exposure)
 # Ensure exposure_df is named and other necessary parameters like activity_type are provided
 p <- plot_exposure(exposure_df = athlytics_sample_exposure, activity_type = "Run")
 #> Generating plot...
