### 2026-01-03 4a78778
fix: add pre-built vignettes to inst/doc for R CMD check
