+++ inst/doc/athlytics_introduction.R
@@ -1,0 +1,367 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7,
+  fig.height = 5
+)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # CRAN (stable)
+# install.packages("Athlytics")
+# 
+# # GitHub (latest features)
+# remotes::install_github('HzaCode/Athlytics')
+
+## ----eval=FALSE---------------------------------------------------------------
+# library(Athlytics)
+# library(dplyr)  # For data manipulation
+# 
+# # Load your activities
+# activities <- load_local_activities("path/to/export_12345678.zip")
+# 
+# # View the first few rows
+# head(activities)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # How many activities do you have?
+# nrow(activities)
+# # Example output: [1] 847
+# 
+# # What sports are in your data?
+# table(activities$type)
+# # Example output:
+# #   Ride  Run  Swim
+# #    312  498    37
+# 
+# # Date range
+# range(activities$date, na.rm = TRUE)
+# # Example output: [1] "2018-01-05" "2024-12-20"
+# 
+# # Key columns in the dataset
+# names(activities)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Summary statistics
+# summary(activities %>% select(distance_km, duration_mins, avg_hr))
+# 
+# # Check for missing heart rate data
+# sum(!is.na(activities$avg_hr)) / nrow(activities) * 100
+# # Shows % of activities with HR data
+# 
+# # Activities without HR data
+# activities %>%
+#   filter(is.na(avg_hr)) %>%
+#   count(sport)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Only running activities
+# runs <- activities %>%
+#   filter(sport == "Run")
+# 
+# # Recent activities (last 6 months)
+# recent <- activities %>%
+#   filter(date >= Sys.Date() - 180)
+# 
+# # Runs with heart rate data from 2024
+# runs_2024_hr <- activities %>%
+#   filter(sport == "Run",
+#          !is.na(avg_hr),
+#          lubridate::year(date) == 2024)
+# 
+# # Long runs only (> 15 km)
+# long_runs <- activities %>%
+#   filter(sport == "Run", distance_km > 15)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Calculate ACWR for all running activities
+# acwr_data <- calculate_acwr(
+#   activities_data = runs,
+#   activity_type = "Run",        # Filter by sport
+#   load_metric = "duration_mins", # Can also be "distance_km" or "hrss"
+#   acute_period = 7,              # 7-day rolling average
+#   chronic_period = 28            # 28-day rolling average
+# )
+# 
+# # View results
+# head(acwr_data)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Basic plot
+# plot_acwr(acwr_data)
+# 
+# # With risk zones highlighted (recommended)
+# plot_acwr(acwr_data, highlight_zones = TRUE)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Identify high-risk periods
+# high_risk <- acwr_data %>%
+#   filter(acwr_smooth > 1.5) %>%
+#   select(date, daily_load, acwr_smooth)
+# 
+# print(high_risk)
+# 
[…]
