+++ inst/doc/athlytics_introduction.Rmd
@@ -1,0 +1,714 @@
+---
+title: "Getting Started with Athlytics"
+author: "Athlytics Package"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Getting Started with Athlytics}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7,
+  fig.height = 5
+)
+```
+
+# Welcome to Athlytics
+
+This tutorial will guide you through a complete workflow—from loading your Strava data to generating training analytics. By the end, you'll understand how to use all core features of Athlytics for longitudinal exercise physiology analysis.
+
+**What You'll Learn:**
+
+- How to load and explore your Strava export data
+- Calculate and interpret training load metrics (ACWR)
+- Analyze aerobic fitness trends (Efficiency Factor)
+- Quantify cardiovascular drift (Decoupling)
+- Track personal bests and performance progression
+- Export results for further analysis
+
+**Time Required:** 30-60 minutes
+
+---
+
+# Prerequisites
+
+## Installation
+
+For installation instructions, see the [README](https://github.com/HzaCode/Athlytics#installation). The quick version:
+
+```{r eval=FALSE}
+# CRAN (stable)
+install.packages("Athlytics")
+
+# GitHub (latest features)
+remotes::install_github('HzaCode/Athlytics')
+```
+
+## Your Strava Data Export
+
+You'll need a Strava data export ZIP file. If you haven't exported your data yet, see the [Strava Export Guide](https://support.strava.com/hc/en-us/articles/216918437-Exporting-your-Data-and-Bulk-Export) or follow the steps in the [README Quick Start](https://github.com/HzaCode/Athlytics#quick-start).
+
+**Quick Summary:**
+1. Go to Strava Settings → My Account → Download or Delete Your Account
+2. Request "Download" (NOT delete!)
+3. Wait for email with download link
+4. Download the ZIP file (e.g., `export_12345678.zip`)
+5. **Don't unzip it** — Athlytics reads ZIP files directly
+
+---
+
+# Loading Your Data
+
+## Basic Loading
+
+```{r eval=FALSE}
+library(Athlytics)
+library(dplyr)  # For data manipulation
+
+# Load your activities
+activities <- load_local_activities("path/to/export_12345678.zip")
+
+# View the first few rows
+head(activities)
+```
+
+## Understanding the Data Structure
+
+Let's explore what we just loaded:
+
+```{r eval=FALSE}
+# How many activities do you have?
+nrow(activities)
+# Example output: [1] 847
+
+# What sports are in your data?
+table(activities$type)
+# Example output:
+#   Ride  Run  Swim
+#    312  498    37
+
+# Date range
+range(activities$date, na.rm = TRUE)
+# Example output: [1] "2018-01-05" "2024-12-20"
+
+# Key columns in the dataset
+names(activities)
+```
+
+**Important Columns:**
+
+- `activity_id` — Unique identifier
+- `date` — Activity date
+- `sport` — Activity type (Run, Ride, Swim, etc.)
+- `distance_km` — Distance in kilometers
+- `duration_mins` — Duration in minutes
+- `avg_hr` — Average heart rate (if recorded)
+- `max_hr` — Maximum heart rate
[…]

[…]
