+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   advanced_features: advanced_features.html
   athlytics_introduction: athlytics_introduction.html
-last_built: 2026-01-03T08:16Z
+last_built: 2026-01-03T11:47Z
 urls:
   reference: https://hzacode.github.io/Athlytics/reference
   article: https://hzacode.github.io/Athlytics/articles
