### 2026-01-04 360335b
ci: use official rOpenSci pkgcheck config
