### 2026-01-16 91e4979
v1.0.2: runnable vignettes, sample datasets, styler formatting
