+++ R/load_local_activities.R
@@ -70,7 +70,7 @@
 #' # Use with Athlytics functions
 #' acwr_data <- calculate_acwr(activities, load_metric = "distance_km")
 #' plot_acwr(acwr_data, highlight_zones = TRUE)
-#' 
+#'
 #' # Multi-metric analysis
 #' ef_data <- calculate_ef(activities, ef_metric = "pace_hr")
 #' plot_ef(ef_data, add_trend_line = TRUE)
@@ -82,64 +82,72 @@
 #' @importFrom rlang .data
 #' @export
 load_local_activities <- function(path = "strava_export_data/activities.csv",
-                                 start_date = NULL,
-                                 end_date = NULL,
-                                 activity_types = NULL) {
-  
+                                  start_date = NULL,
+                                  end_date = NULL,
+                                  activity_types = NULL) {
   # --- Input Validation ---
   if (!file.exists(path)) {
-    stop("File not found at: ", path,
-         "\nPlease ensure you have downloaded your Strava data export.")
+    stop(
+      "File not found at: ", path,
+      "\nPlease ensure you have downloaded your Strava data export."
+    )
   }
-  
+
   # --- Handle ZIP files ---
   temp_extracted <- FALSE
   original_path <- path
-  
+
   # Check if input is a ZIP file
   if (tolower(tools::file_ext(path)) == "zip") {
     message("Detected ZIP archive. Extracting activities.csv...")
-    
+
     # Create a temporary directory for extraction
     temp_dir <- tempdir()
-    
+
     # List files in the ZIP
     zip_contents <- utils::unzip(path, list = TRUE)
-    
+
     # Find activities.csv (case-insensitive)
-    activities_file <- grep("activities\\.csv$", zip_contents$Name, 
-                           ignore.case = TRUE, value = TRUE)
-    
+    activities_file <- grep("activities\\.csv$", zip_contents$Name,
+      ignore.case = TRUE, value = TRUE
+    )
+
     if (length(activities_file) == 0) {
       stop("No activities.csv file found in ZIP archive: ", path)
     }
-    
+
     if (length(activities_file) > 1) {
       warning("Multiple activities.csv files found. Using: ", activities_file[1])
       activities_file <- activities_file[1]
     }
-    
+
     # Extract the activities.csv file
-    tryCatch({
-      utils::unzip(path, files = activities_file, exdir = temp_dir, overwrite = TRUE)
-      path <- file.path(temp_dir, activities_file)
-      temp_extracted <- TRUE
-      message("Successfully extracted to temporary location.")
-    }, error = function(e) {
-      stop("Failed to extract ZIP archive: ", e$message)
-    })
+    tryCatch(
+      {
+        utils::unzip(path, files = activities_file, exdir = temp_dir, overwrite = TRUE)
+        path <- file.path(temp_dir, activities_file)
+        temp_extracted <- TRUE
+        message("Successfully extracted to temporary location.")
+      },
+      error = function(e) {
+        stop("Failed to extract ZIP archive: ", e$message)
+      }
+    )
   }
-  
+
   # --- Read CSV ---
   message("Reading activities from: ", basename(original_path))
-  
+
   # Read CSV with appropriate column types
-  activities_raw <- tryCatch({
-    readr::read_csv(path, show_col_types = FALSE, col_types = readr::cols())
-  }, error = function(e) {
-    stop("Failed to read CSV file: ", e$message)
-  })
-  
+  activities_raw <- tryCatch(
+    {
+      readr::read_csv(path, show_col_types = FALSE, col_types = readr::cols())
+    },
[…]
