+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(Athlytics)
 
-test_check("Athlytics") 
+test_check("Athlytics")
