+++ reference/calculate_cohort_reference.md
@@ -1,0 +1,166 @@
+# Calculate Cohort Reference Percentiles
+
+Calculates reference percentiles for a metric across a cohort of
+athletes, stratified by specified grouping variables (e.g., sport, sex,
+age band).
+
+## Usage
+
+``` r
+calculate_cohort_reference(
+  data,
+  metric = "acwr_smooth",
+  by = c("sport"),
+  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
+  min_athletes = 5,
+  date_col = "date"
+)
+
+cohort_reference(
+  data,
+  metric = "acwr_smooth",
+  by = c("sport"),
+  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
+  min_athletes = 5,
+  date_col = "date"
+)
+```
+
+## Arguments
+
+- data:
+
+  A data frame containing metric values for multiple athletes. Must
+  include columns: `date`, `athlete_id`, and the metric column.
+
+- metric:
+
+  Name of the metric column to calculate percentiles for (e.g., "acwr",
+  "acwr_smooth", "ef", "decoupling"). Default "acwr_smooth".
+
+- by:
+
+  Character vector of grouping variables. Options: "sport", "sex",
+  "age_band", "athlete_id". Default c("sport").
+
+- probs:
+
+  Numeric vector of probabilities for percentiles (0-1). Default c(0.05,
+  0.25, 0.50, 0.75, 0.95) for 5th, 25th, 50th, 75th, 95th percentiles.
+
+- min_athletes:
+
+  Minimum number of athletes required per group to calculate valid
+  percentiles. Default 5.
+
+- date_col:
+
+  Name of the date column. Default "date".
+
+## Value
+
+A long-format data frame with columns:
+
+- date:
+
+  Date
+
+- ...:
+
+  Grouping variables (as specified in `by`)
+
+- percentile:
+
+  Percentile label (e.g., "p05", "p25", "p50", "p75", "p95")
+
+- value:
+
+  Metric value at that percentile
+
+- n_athletes:
+
+  Number of athletes contributing to this percentile
+
+## Details
+
+This function creates cohort-level reference bands for comparing
+individual athlete metrics to their peers. Common use cases:
+
+- Compare an athlete's ACWR trend to team averages
+
+- Identify outliers (athletes outside P5-P95 range)
+
+- Track team-wide trends over time
+
+**Important**: Percentile bands represent **population variability**,
+not statistical confidence intervals for individual values.
+
+## Examples
+
+``` r
+# Example using sample data to create a mock cohort
+data("sample_acwr", package = "Athlytics")
+
+# Simulate a cohort by duplicating with different athlete IDs
+cohort_mock <- dplyr::bind_rows(
+  dplyr::mutate(sample_acwr, athlete_id = "A1", sport = "Run"),
+  dplyr::mutate(sample_acwr,
+    athlete_id = "A2", sport = "Run",
+    acwr_smooth = acwr_smooth * runif(nrow(sample_acwr), 0.9, 1.1)
+  ),
+  dplyr::mutate(sample_acwr,
+    athlete_id = "A3", sport = "Run",
+    acwr_smooth = acwr_smooth * runif(nrow(sample_acwr), 0.85, 1.15)
+  )
+)
+
+# Calculate reference percentiles (min_athletes = 2 for demo)
+reference <- calculate_cohort_reference(cohort_mock,
+  metric = "acwr_smooth",
+  by = "sport", min_athletes = 2
+)
+#> Calculating percentiles for metric 'acwr_smooth' grouped by: sport
+#> Reference calculated for 332 date-group combinations.
[…]
