### 2026-01-16 8c0af27
ci: bust cache to fix rlang version conflict
