### 2026-01-17 aec3345
ci: fix pkgcheck rlang version conflict
