+++ .github/workflows/pkgcheck.yml
@@ -15,7 +15,22 @@ jobs:
     runs-on: ubuntu-latest
     permissions:
       issues: write
+    env:
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
-      - uses: ropensci-review-tools/pkgcheck-action@main
+      - uses: actions/checkout@v4
+      
+      - uses: r-lib/actions/setup-r@v2
+        with:
+          use-public-rspm: true
+      
+      - name: Force update rlang to fix version conflict
+        run: |
+          install.packages("rlang", repos = "https://cloud.r-project.org")
+        shell: Rscript {0}
+      
+      - uses: ropensci-review-tools/pkgcheck-action@v1
+        with:
+          post-to-issue: false
 
 
