### 2026-01-18 d38f8cd
docs: add CITATION.cff and sync citation info to v1.0.2
