+++ index.md
@@ -193,7 +193,7 @@ well as the original methodological sources for specific metrics.
   title   = {Athlytics: A Reproducible Framework for Endurance Data Analysis},
   author  = {Zhiang He},
   year    = {2025},
-  version = {1.0.0},
+  version = {1.0.2},
   url     = {https://github.com/HzaCode/Athlytics}
 }
 ```
