### 2026-01-27 #9075 @HzaCode
Hi @rkurchin and @juliaferraioli,

Thank you for the check-in.

The review process at rOpenSci is progressing smoothly. I am currently addressing the reviewers' feedback and finalizing the revisions.

I will update this thread immediately once the rOpenSci review concludes.
