### 2026-02-07 f52d6aa
v1.0.3: address reviewer feedback with bug fixes, test rewrite, and documentation improvements
## Test Suite Overhaul
- Rewrite test suite from ~1500 bloated tests to ~200 focused, meaningful tests
- Replace redundant and overlapping test files with consolidated test modules
- Add concrete value assertions (expect_equal with tolerance) instead of trivial smoke tests
- Remove 15 redundant test files: test-absolute-real-data.R, test-acwr-ewma.R,
  test-acwr-ewma-advanced.R, test-calculate-ef-extended.R, test-calculate-ef-simple.R,
  test-calculate-ef-stream.R, test-calculate-exposure-extended.R,
[…]
