+++ tests/testthat/test-plot-ef-advanced.R
@@ -1,222 +1,0 @@
-# Advanced test for plot_ef.R to boost coverage further
-
-test_that("plot_ef handles complex data scenarios", {
-  # Create mock activities data with more complex scenarios
-  mock_activities <- data.frame(
-    date = seq(Sys.Date() - 200, Sys.Date(), by = "3 days"),
-    type = rep("Run", 67),
-    moving_time = rep(2400, 67),
-    distance = rep(8000, 67),
-    average_heartrate = seq(140, 160, length.out = 67),
-    average_watts = rep(200, 67),
-    filename = rep(NA, 67),
-    stringsAsFactors = FALSE
-  )
-
-  # Test plot_ef with complex data calculation
-  p <- plot_ef(
-    data = mock_activities,
-    activity_type = "Run",
-    ef_metric = "pace_hr",
-    start_date = Sys.Date() - 100,
-    end_date = Sys.Date(),
-    min_duration_mins = 15
-  )
-  expect_s3_class(p, "ggplot")
-
-  # Test with different smoothing parameters
-  p2 <- plot_ef(
-    data = mock_activities,
-    activity_type = "Run",
-    ef_metric = "pace_hr",
-    smoothing_method = "loess",
-    add_trend_line = TRUE
-  )
-  expect_s3_class(p2, "ggplot")
-})
-
-test_that("plot_ef handles edge cases with data", {
-  # Test with data that has gaps
-  mock_gapped <- data.frame(
-    date = c(Sys.Date() - 100, Sys.Date() - 50, Sys.Date() - 10, Sys.Date()),
-    type = rep("Run", 4),
-    moving_time = rep(2400, 4),
-    distance = rep(8000, 4),
-    average_heartrate = c(150, 155, 160, 165),
-    average_watts = rep(200, 4),
-    filename = rep(NA, 4),
-    stringsAsFactors = FALSE
-  )
-
-  p <- plot_ef(
-    data = mock_gapped,
-    activity_type = "Run",
-    ef_metric = "pace_hr"
-  )
-  expect_s3_class(p, "ggplot")
-
-  # Test with data that has outliers
-  mock_outliers <- data.frame(
-    date = seq(Sys.Date() - 50, Sys.Date(), by = "7 days"),
-    type = rep("Run", 8),
-    moving_time = rep(2400, 8),
-    distance = rep(8000, 8),
-    average_heartrate = c(150, 155, 160, 200, 165, 170, 175, 180), # Outlier at 200
-    average_watts = rep(200, 8),
-    filename = rep(NA, 8),
-    stringsAsFactors = FALSE
-  )
-
-  p2 <- plot_ef(
-    data = mock_outliers,
-    activity_type = "Run",
-    ef_metric = "pace_hr"
-  )
-  expect_s3_class(p2, "ggplot")
-})
-
-test_that("plot_ef handles different smoothing scenarios", {
-  data("sample_ef")
-
-  # Test with different smoothing methods and parameters
-  p1 <- plot_ef(sample_ef, smoothing_method = "loess", add_trend_line = TRUE)
-  expect_s3_class(p1, "ggplot")
-
-  p2 <- plot_ef(sample_ef, smoothing_method = "lm", add_trend_line = FALSE)
-  expect_s3_class(p2, "ggplot")
-
-  p3 <- plot_ef(sample_ef, smoothing_method = "gam", add_trend_line = TRUE)
-  expect_s3_class(p3, "ggplot")
-})
-
-test_that("plot_ef handles activity type filtering edge cases", {
-  data("sample_ef")
-
[…]
