+++ R/cohort_reference.R
@@ -29,19 +29,13 @@
 #' @details
 #' This function creates cohort-level reference bands for comparing individual
 #' athlete metrics to their peers. Common use cases:
-#' \itemize{
-#'   \item Compare an athlete's ACWR trend to team averages
-#'   \item Identify outliers (athletes outside P5-P95 range)
-#'   \item Track team-wide trends over time
-#' }
+#' - Compare an athlete's ACWR trend to team averages
+#' - Identify outliers (athletes outside P5-P95 range)
+#' - Track team-wide trends over time
 #'
 #' **Important**: Percentile bands represent **population variability**, not
 #' statistical confidence intervals for individual values.
 #'
-#' @importFrom dplyr group_by summarise filter mutate %>% ungroup
-#' @importFrom tidyr pivot_longer
-#' @importFrom stats quantile setNames
-#' @importFrom rlang .data
 #' @export
 #'
 #' @examples
@@ -139,7 +133,7 @@ calculate_cohort_reference <- function(data,
   }
 
   # --- Calculate Percentiles ---
-  message(sprintf(
+  athlytics_message(sprintf(
     "Calculating percentiles for metric '%s' grouped by: %s",
     metric, paste(by, collapse = ", ")
   ))
@@ -185,7 +179,7 @@ calculate_cohort_reference <- function(data,
       values_to = "value"
     )
 
-  message(sprintf(
+  athlytics_message(sprintf(
     "Reference calculated for %d date-group combinations.",
     nrow(reference_data)
   ))
@@ -228,9 +222,6 @@ cohort_reference <- function(data,
 #'
 #' @return A ggplot object with added reference bands.
 #'
-#' @importFrom ggplot2 geom_ribbon geom_line aes
-#' @importFrom dplyr filter %>%
-#' @importFrom tidyr pivot_wider
 #' @export
 #'
 #' @examples
@@ -247,9 +238,9 @@ add_reference_bands <- function(p,
                                 bands = c("p25_p75", "p05_p95", "p50"),
                                 alpha = 0.15,
                                 colors = list(
-                                  p25_p75 = "#440154FF", # viridis dark purple
-                                  p05_p95 = "#3B528BFF", # viridis blue
-                                  p50 = "#21908CFF" # viridis teal
+                                  p25_p75 = "#4DBBD5",
+                                  p05_p95 = "#E64B35",
+                                  p50 = "#3C5488"
                                 )) {
   if (!inherits(p, "ggplot")) {
     stop("`p` must be a ggplot object.")
@@ -261,8 +252,8 @@ add_reference_bands <- function(p,
 
   # Pivot reference data to wide format for plotting
   ref_wide <- reference_data %>%
-    dplyr::select(.data$date, .data$percentile, .data$value) %>%
-    tidyr::pivot_wider(names_from = .data$percentile, values_from = .data$value)
+    dplyr::select("date", "percentile", "value") %>%
+    tidyr::pivot_wider(names_from = "percentile", values_from = "value")
 
   # Add bands in order (outermost to innermost)
   if ("p05_p95" %in% bands && all(c("p05", "p95") %in% colnames(ref_wide))) {
@@ -308,31 +299,30 @@ add_reference_bands <- function(p,
 #' @param date_col Name of the date column. Default "date".
 #' @param title Plot title. Default NULL (auto-generated).
[…]
