### 2026-02-07 4676d2c
ci: fix FITfileR dependency resolution in GitHub Actions
Add github::grimbough/FITfileR to extra-packages in both R-CMD-check and pkgdown workflows so pak can resolve the GitHub-only dependency. Remove redundant manual install step from R-CMD-check.
