+++ .github/workflows/pkgdown.yml
@@ -30,7 +30,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::pkgdown, local::.
+          extra-packages: any::pkgdown, local::., github::grimbough/FITfileR
           needs: website
 
       - name: Build site
