### 2026-02-07 10906dd
ci: fix test-acwr floating-point tolerance and enable vignette build in CI
Fix ACWR test failure on ubuntu-latest by adding set.seed(42) for reproducibility and relaxing ATL non-negative check to allow floating-point tolerance. Enable vignette building in R CMD build/check to resolve inst/doc warnings.
