+++ reference/calculate_acwr.md
@@ -29,11 +29,12 @@ calculate_acwr(
   acute_period = 7,
   chronic_period = 28,
   start_date = NULL,
-  end_date = NULL,
+  end_date = Sys.Date(),
   user_ftp = NULL,
   user_max_hr = NULL,
   user_resting_hr = NULL,
-  smoothing_period = 7
+  smoothing_period = 7,
+  verbose = FALSE
 )
 ```
 
@@ -113,6 +114,10 @@ calculate_acwr(
   Integer. Number of days for smoothing the ACWR using a rolling mean
   (default: 7). Reduces day-to-day noise for clearer trend
   visualization.
+
+- verbose:
+
+  Logical. If TRUE, prints progress messages. Default FALSE.
 
 ## Value
 
@@ -199,25 +204,50 @@ column before calculation and use `group_by(athlete_id)` with
 [`group_modify()`](https://dplyr.tidyverse.org/reference/group_map.html).
 See examples below and vignettes for details.
 
+## Scientific Considerations
+
+**Important**: The predictive value of ACWR for injury risk has been
+debated in recent literature. Some researchers argue that ACWR may have
+limited utility for predicting injuries (Impellizzeri et al., 2020), and
+a subsequent analysis has called for dismissing the ACWR framework
+entirely (Impellizzeri et al., 2021). Users should interpret ACWR risk
+zones with caution and consider them as descriptive heuristics rather
+than validated injury predictors.
+
+Impellizzeri, F. M., Tenan, M. S., Kempton, T., Novak, A., & Coutts, A.
+J. (2020). Acute:chronic workload ratio: conceptual issues and
+fundamental pitfalls. *International Journal of Sports Physiology and
+Performance*, 15(6), 907-913.
+[doi:10.1123/ijspp.2019-0864](https://doi.org/10.1123/ijspp.2019-0864)
+
+Impellizzeri, F. M., Woodcock, S., Coutts, A. J., Fanchini, M., McCall,
+A., & Vigotsky, A. D. (2021). What role do chronic workloads play in the
+acute to chronic workload ratio? Time to dismiss ACWR and its underlying
+theory. *Sports Medicine*, 51(3), 581-592.
+[doi:10.1007/s40279-020-01378-6](https://doi.org/10.1007/s40279-020-01378-6)
+
 ## References
 
 Gabbett, T. J. (2016). The training-injury prevention paradox: should
 athletes be training smarter and harder? *British Journal of Sports
 Medicine*, 50(5), 273-280.
+[doi:10.1136/bjsports-2015-095788](https://doi.org/10.1136/bjsports-2015-095788)
 
-Hulin, B. T., et al. (2016). The acute:chronic workload ratio predicts
-injury: high chronic workload may decrease injury risk in elite rugby
-league players. *British Journal of Sports Medicine*, 50(4), 231-236.
+Hulin, B. T., Gabbett, T. J., Lawson, D. W., Caputi, P., & Sampson, J.
+A. (2016). The acute:chronic workload ratio predicts injury: high
+chronic workload may decrease injury risk in elite rugby league players.
+*British Journal of Sports Medicine*, 50(4), 231-236.
+[doi:10.1136/bjsports-2015-094817](https://doi.org/10.1136/bjsports-2015-094817)
 
 ## See also
 
-[`plot_acwr`](https://hzacode.github.io/Athlytics/reference/plot_acwr.md)
+[`plot_acwr()`](https://hzacode.github.io/Athlytics/reference/plot_acwr.md)
 for visualization,
[…]
