+++ reference/calculate_pbs.md
@@ -12,8 +12,9 @@ calculate_pbs(
   export_dir = "strava_export_data",
   activity_type = "Run",
   start_date = NULL,
-  end_date = NULL,
-  distances_m = c(1000, 5000, 10000, 21097.5, 42195)
+  end_date = Sys.Date(),
+  distances_m = c(1000, 5000, 10000, 21097.5, 42195),
+  verbose = FALSE
 )
 ```
 
@@ -36,18 +37,22 @@ calculate_pbs(
 
 - start_date:
 
-  Optional start date for analysis (YYYY-MM-DD). Defaults to NULL (all
-  dates).
+  Optional start date for analysis (YYYY-MM-DD). If NULL, defaults to
+  365 days before `end_date`.
 
 - end_date:
 
-  Optional end date for analysis (YYYY-MM-DD). Defaults to NULL (all
-  dates).
+  End date for analysis (YYYY-MM-DD). Default
+  [`Sys.Date()`](https://rdrr.io/r/base/Sys.time.html) (today).
 
 - distances_m:
 
   Target distances in meters to track. Default: c(1000, 5000, 10000,
   21097.5, 42195) for 1k, 5k, 10k, half, full marathon.
+
+- verbose:
+
+  Logical. If TRUE, prints progress messages. Default FALSE.
 
 ## Value
 
@@ -61,6 +66,18 @@ This function analyzes detailed activity files (FIT/TCX/GPX) to find the
 fastest efforts at specified distances. It tracks cumulative personal
 bests over time, showing when new PBs are set.
 
+Personal best tracking is a standard approach in endurance sport
+performance monitoring. Systematic PB analysis over multiple distances
+helps identify fitness improvements, training phase effectiveness, and
+performance peaks (Matveyev, 1981). The multi-distance approach enables
+athletes to assess both speed (shorter distances) and endurance (longer
+distances) progression simultaneously.
+
+### References
+
+- Matveyev, L. P. (1981). *Fundamentals of Sports Training*. Moscow:
+  Progress Publishers.
+
 **Note**: Requires detailed activity files from your Strava export.
 Activities must be long enough to contain the target distance segments.
 
@@ -73,12 +90,12 @@ print(head(sample_pbs))
 #> # A tibble: 6 × 10
 #>   activity_id activity_date       distance elapsed_time moving_time time_seconds
 #>   <chr>       <dttm>                 <dbl>        <dbl>       <dbl>        <dbl>
-#> 1 activity_1  2023-01-01 00:00:00     1000         307.        307.         307.
-#> 2 activity_2  2023-01-01 00:00:00     5000        1479.       1479.        1479.
-#> 3 activity_3  2023-01-01 00:00:00    10000        3047.       3047.        3047.
-#> 4 activity_4  2023-02-01 00:00:00     1000         293.        293.         293.
-#> 5 activity_5  2023-03-01 00:00:00     1000         294.        294.         294.
-#> 6 activity_6  2023-03-01 00:00:00     5000        1451.       1451.        1451.
+#> 1 activity_1  2023-01-01 00:00:00    1000           256         256          256
+#> 2 activity_2  2023-01-01 00:00:00    5000          1373        1373         1373
+#> 3 activity_3  2023-01-01 00:00:00   10000          2877        2877         2877
+#> 4 activity_4  2023-01-01 00:00:00   21098.         6582        6582         6582
+#> 5 activity_5  2023-02-01 00:00:00    1000           255         255          255
+#> 6 activity_6  2023-03-01 00:00:00    1000           246         246          246
[…]
