### 2026-02-07 6920ffc
ci: add Remotes and Additional_repositories for FITfileR dependency
