### 2026-02-13 66c8086
address smnnlt follow-up: streamline test suite and drop purrr
- rewrote test suite from ~600 to 373 focused assertions (0 fail / 0 warn / 0 skip on local devtools::test())

- replaced expect_true(is.data.frame()) with expect_s3_class(), expect_equal(length()) with expect_length()

- removed redundant gg-class checks now covered by vdiffr snapshots

- deleted test-smoke-and-errors.R, test-uncovered-branches.R and other fragmented test files

- replaced mock helpers with inst/extdata/ example data where possible

- removed purrr from dependencies entirely

- removed CITATION.cff (flagged by R C
[…]
