+++ tests/testthat/test-pbs.R
@@ -1,127 +1,48 @@
 # tests/testthat/test-pbs.R
+# PBS calculation and plotting tests
 
-library(testthat)
-library(Athlytics)
-suppressWarnings(library(ggplot2)) # Explicitly load for s3_class checks if not automatically available
-library(lubridate) # For seconds_to_period if used in manual_df
+# --- plot_pbs using sample_pbs ---
 
-# Load data: sample data from package & helper mock data
-data(sample_pbs)
-source(test_path("helper-mockdata.R"), local = TRUE)
+test_that("plot_pbs produces expected layers with sample_pbs", {
+  data("sample_pbs")
 
-# --- Test plot_pbs using manually created, well-structured pbs_df ---
-
-test_that("plot_pbs returns a ggplot object with valid pbs_df input", {
-  # 1. Manually create a small, representative pbs_df
-  #    This data frame should have the exact column names and types
-  #    that calculate_pbs would output and plot_pbs expects.
-  manual_pbs_df <- data.frame(
-    activity_date = as.Date(c("2023-01-01", "2023-01-08", "2023-01-15", "2023-01-01", "2023-01-20")),
-    time_seconds = c(1200, 1180, 1190, 300, 1170), # Example times
-    distance = c(5000, 5000, 5000, 1000, 5000), # Example distances
-    is_pb = c(TRUE, TRUE, FALSE, TRUE, TRUE), # Example PB flags
-    distance_label = factor(c("5k", "5k", "5k", "1k", "5k"), levels = c("1k", "5k")),
-    activity_id = as.character(1:5),
-    elapsed_time = c(1200, 1180, 1190, 300, 1170),
-    moving_time = c(1200, 1180, 1190, 300, 1170),
-    cumulative_pb_seconds = c(1200, 1180, 1180, 300, 1170),
-    # time_period is derived by plot_pbs if not present, or by calculate_pbs
-    # For a direct pbs_df input test, ensure it has what plot_pbs *uses* or can work without.
-    # calculate_pbs does output time_period.
-    stringsAsFactors = FALSE
-  )
-  # Add time_period for completeness matching calculate_pbs output
-  manual_pbs_df$time_period <- lubridate::seconds_to_period(manual_pbs_df$time_seconds)
-
-
-  # Ensure distance_meters argument for plot_pbs matches distances in manual_pbs_df
-  test_distance_meters <- unique(manual_pbs_df$distance)
-
-  # 2. Call plot_pbs with this manually created data
-  # suppressWarnings: distance_meters is analysis arg, ignored with pre-calculated data
-  p <- suppressWarnings(plot_pbs(data = manual_pbs_df, distance_meters = test_distance_meters))
-
-  # 3. Perform assertions on the plot object 'p'
-  expect_s3_class(p, "ggplot")
-
-  # Check for expected layers if possible (e.g., points, lines)
-  expect_true(length(p$layers) >= 2) # Expect at least geom_line and geom_point
-
-  # Check labels (example)
-  expect_equal(p$labels$x, "Activity Date")
-  expect_equal(p$labels$y, "Best Time (MM:SS)")
-  expect_equal(p$labels$title, "Personal Best Running Times Trend")
+  p <- plot_pbs(data = sample_pbs)
+  expect_gte(length(p$layers), 2) # At least geom_line and geom_point
 })
 
-test_that("plot_pbs handles empty data frame input", {
-  # Create an empty data frame with the correct column structure
+test_that("plot_pbs handles empty data frame", {
   empty_df <- data.frame(
-    activity_date = as.Date(character()),
-    time_seconds = numeric(),
[…]
