### 2026-02-13 ff7eb3f
bump version to 1.0.4
