+++ authors.md
@@ -10,12 +10,12 @@ Source:
 [`inst/CITATION`](https://github.com/HzaCode/Athlytics/blob/main/inst/CITATION)
 
 Zhiang He (2025). Athlytics: A Reproducible Framework for Endurance Data
-Analysis. R package version 1.0.3. https://github.com/HzaCode/Athlytics
+Analysis. R package version 1.0.4. https://github.com/HzaCode/Athlytics
 
     @Manual{,
       title = {Athlytics: A Reproducible Framework for Endurance Data Analysis},
       author = {Zhiang He},
       year = {2025},
-      note = {R package version 1.0.3},
+      note = {R package version 1.0.4},
       url = {https://github.com/HzaCode/Athlytics},
     }
