### 2026-02-13 f292b28
fix: make add_reference_bands example self-contained to fix pkgcheck
