### 2026-02-19 e6637a7
docs: restore CITATION.cff and update to v1.0.4
