+++ CITATION.cff
@@ -1,0 +1,21 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+type: software
+title: "Athlytics: A Reproducible Framework for Endurance Data Analysis"
+version: 1.0.4
+date-released: 2025-02-19
+url: "https://github.com/HzaCode/Athlytics"
+repository-code: "https://github.com/HzaCode/Athlytics"
+license: MIT
+authors:
+  - family-names: "He"
+    given-names: "Zhiang"
+    email: "ang@hezhiang.com"
+keywords:
+  - R
+  - sports-science
+  - exercise-physiology
+  - training-load
+  - ACWR
+  - Strava
+abstract: "An open-source computational framework for longitudinal analysis of exercise physiology metrics using local Strava data exports. Provides standardized functions to calculate and visualize key physiological indicators including Acute:Chronic Workload Ratio (ACWR), Efficiency Factor (EF), and training load metrics."
