+++ reference/calculate_exposure.md
@@ -125,12 +125,12 @@ exposure_result <- calculate_exposure(
 )
 print(head(exposure_result))
 #>         date daily_load atl ctl acwr
-#> 1 2026-01-17         10  10  10    1
-#> 2 2026-01-18         10  10  10    1
-#> 3 2026-01-19         10  10  10    1
-#> 4 2026-01-20         10  10  10    1
-#> 5 2026-01-21         10  10  10    1
-#> 6 2026-01-22         10  10  10    1
+#> 1 2026-01-23         10  10  10    1
+#> 2 2026-01-24         10  10  10    1
+#> 3 2026-01-25         10  10  10    1
+#> 4 2026-01-26         10  10  10    1
+#> 5 2026-01-27         10  10  10    1
+#> 6 2026-01-28         10  10  10    1
 
 if (FALSE) { # \dontrun{
 # Example using local Strava export data
