### 2026-02-19 12c6bbf
docs: enhance paper.md with expanded Software Description, reproducible example, and updated acknowledgements
