+++ paper/paper.md
@@ -18,7 +18,7 @@ affiliations:
     index: 1
 date: "2025-10-11"
 bibliography: paper.bib
-version: 1.0.0
+version: 1.0.4
 license: MIT
 ---
 
@@ -28,7 +28,7 @@ license: MIT
 
 # Statement of Need
 
-Analyzing endurance training data in R often requires stitching together API clients, file parsers, and custom scripts. This creates workflows that are fragile and difficult to reproduce, especially for **cohort-scale** research. **Athlytics** addresses this gap by providing a **single, research-oriented pipeline** that works offline with local archives, offers an integrated suite of physiological models, and is built from the ground up for multi-athlete analysis. This design **reduces "glue code"** and makes cohort-scale analyses auditable and easy to reproduce.
+Analyzing endurance training data in R often requires stitching together API clients, file parsers, and custom scripts. This creates workflows that are fragile and difficult to reproduce, especially for **cohort-scale** research. **Athlytics** addresses this gap by providing a **single, research-oriented pipeline** that works offline with local archives, offers an integrated suite of physiological models, and is built from the ground up for multi-athlete analysis. This design **reduces "glue code"** and makes cohort-scale analyses auditable and easy to reproduce. The primary audience for Athlytics includes **sports scientists, sports epidemiologists, and endurance coaches** who need a reliable, programmatic way to process cohort-scale Strava data reproducibly, bridging the gap between raw XML/FIT files and advanced physiological models.
 
 # Related Work 
 
@@ -49,11 +49,12 @@ Athlytics is unique in providing an API-free, end-to-end workflow that integrate
 
 # Software Description
 
--   **Inputs & Data Model:** Reads Strava ZIP archives or `activities.csv`. Activity streams (FIT/TCX/GPX) are loaded **on demand** from the archive, optionally using `FITfileR` for FIT files [@FITfileR]. The core function `load_local_activities()` produces a standardized tibble.
--   **Core Metrics:** `calculate_acwr()`, `calculate_ef()`, and `calculate_decoupling()` compute key summaries with sensible, research-oriented defaults. The implementation of ACWR acknowledges its conceptual issues and is presented as a monitoring tool [@impellizzeri2020acwr; @impellizzeri2021dismiss].
--   **Uncertainty Quantification:** Provides confidence intervals for EWMA-based ACWR using a moving-block bootstrap [@kunsch1989; @politis1994], a key feature for research applications.
--   **Cohort Benchmarking:** `calculate_cohort_reference()` computes percentile bands, which can be layered onto individual plots using `plot_with_reference()`.
--   **Plotting & Diagnostics:** Visualization functions follow a **data-first API**. Functions return **diagnostic fields** (e.g., `status`, `reason`) when inputs are insufficient, making the workflow transparent.
+-   **Offline Data Parsing:** Operates directly on local Strava ZIP exports. Using `.fit`, `.tcx`, and `.gpx` parsers via underlying XML and binary decoders (with optional `FITfileR` [@FITfileR] support), activity streams are loaded on demand.
+-   **Physiological & Load Metrics:** Supports multiple load tracking algorithms including HRSS (TRIMP-based) and TSS approximations. Calculates core metrics such as **Aerobic Decoupling**, **Efficiency Factor (EF)**, and automatically tracks **Personal Bests (PBs)** using sliding-window spatial algorithms.
[…]
