### 2026-02-20 1773fd2
docs: update paper.bib reference metadata
