+++ paper/paper.bib
@@ -57,7 +57,7 @@
   title   = {rStrava: Access the Strava API},
   author  = {Marcus W. Beck and Pedro Villarroel and Daniel Padfield and Lorenzo Gaborini and Niklas von Maltzahn},
   year    = {2024},
-  note    = {R package version 1.3.2},
+  note    = {R package version 1.3.4},
   doi     = {10.32614/CRAN.package.rStrava},
   url     = {https://CRAN.R-project.org/package=rStrava}
 }
@@ -78,6 +78,7 @@
   author  = {Daniel Schafer},
   year    = {2024},
   note    = {R package version 0.2.1},
+  doi     = {10.32614/CRAN.package.activatr},
   url     = {https://CRAN.R-project.org/package=activatr}
 }
 
@@ -91,18 +92,19 @@
 }
 
 @Manual{injurytools,
-  title   = {injurytools: A Toolkit for Sports Injury Data Analysis},
+  title   = {injurytools: A Toolkit for Sports Injury and Illness Data Analysis},
   author  = {Lore Zumeta Olaskoaga and Dae-Jin Lee},
   year    = {2023},
-  note    = {R package version 1.0.3},
+  note    = {R package version 2.0.1},
   doi     = {10.32614/CRAN.package.injurytools},
   url     = {https://CRAN.R-project.org/package=injurytools}
 }
 
 @Manual{ACWR,
   title   = {ACWR: Acute Chronic Workload Ratio Calculation},
-  author  = {Ben Torvaney},
+  author  = {Jorge R. Fernandez-Santos},
   year    = {2024},
-  note    = {R package version 0.2.0},
+  note    = {R package version 0.1.0},
+  doi     = {10.32614/CRAN.package.ACWR},
   url     = {https://CRAN.R-project.org/package=ACWR}
 }
